# metrics: add IncCounter benchmark